// src/services/models/upsertUserCreRegistrationDtoWorkflowId.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Reprieve Backend API
 * Debug and integration APIs for positions, rescues, and relay history
 * OpenAPI spec version: 0.1.0
 */

/**
 * Workflow strategy to run for this user
 */
export type UpsertUserCreRegistrationDtoWorkflowId = typeof UpsertUserCreRegistrationDtoWorkflowId[keyof typeof UpsertUserCreRegistrationDtoWorkflowId];


export const UpsertUserCreRegistrationDtoWorkflowId = {
  CHAINLINK_API_GUARD_V1: 'CHAINLINK_API_GUARD_V1',
  QUANT_FUNDING_OI_V1: 'QUANT_FUNDING_OI_V1',
  QUANT_BASIS_LIQUIDITY_V1: 'QUANT_BASIS_LIQUIDITY_V1',
} as const;