// src/services/models/upsertUserCreRegistrationDtoQueuePriority.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * Reprieve Backend API
 * Debug and integration APIs for positions, rescues, and relay history
 * OpenAPI spec version: 0.1.0
 */

/**
 * Rescue priority ordering
 */
export type UpsertUserCreRegistrationDtoQueuePriority = typeof UpsertUserCreRegistrationDtoQueuePriority[keyof typeof UpsertUserCreRegistrationDtoQueuePriority];


export const UpsertUserCreRegistrationDtoQueuePriority = {
  SAME_CHAIN_FIRST: 'SAME_CHAIN_FIRST',
  CROSS_CHAIN_FIRST: 'CROSS_CHAIN_FIRST',
} as const;